Rename finished_books.ts to fetchFinishedBooks.ts
